Move configuration helping methods to ConfigurationFilters
